Wait for Replication and Indexing to complete before becoming leader
- This change ensures that before becoming leader we successfully write an epoch and get it replicated. Thus demonstrating that we are in a fit state to become leader.
- Therefore the log is successfully replicated before we transition to leader.
- We also wait for the records to be indexed before we transition to leader, otherwise incorrect event numbers can be written because the latest events are not known to the index or IndexWriter

- The InaugurationManager is reponsible for making sure the conditions are met for transitioning to leader.
- In the PreLeader state we now accept replication subscriptions and monitor for quorum, transitioning to Unknown if a quorum is not maintained (exactly as we did in the leader state before)

Additionally, whenever we write an epoch we immediately write an $epoch-information event to the $epoch-information stream. This allows the InaugurationManager to know where the
index checkpoint is supposed to reach. As a bonus this also gives the users a way to get details about the epochs. These events will not be present in an unfiltered $all read, for backwards
compatibilty, but can be read directly from their stream.