Prevent deadlock risk when creating a PersistentSubscriptionGroup
The StreamBuffer getter used to implicity block waiting on the _streamBufferSource task.

However, threads that called this would commonly be holding the _lock which OnCheckpointLoaded
needs to acquire to be able to set the result in _streamBufferSource.

Therefore a deadlock could occur.

We have introduced TryGetStreamBuffer which does not block. Caller decides
what to do if the buffer is not available yet.