Use the bloom filter of a scavenged ptable
Previously the bloom filter was not used due to a size mismatch.
This was because the bloom filter is sized according to the pre-scavenge table and populated as the scavenge progresses.
Now the scavenged ptable will use the bloom filter regardless of its size.

Also, delete bloom filter if deleting ptable due to no scavenged entries. This would have been cleaned up on next startup
anyway, but this is tidier.