Fix persistent subscription race condition
A read result of 'NoStream' comes with a NextEventNumber of -1.
If there happens to have been an event written since the read, the PersistentSubscription will issue another read starting from the NextEventNumber
But -1 is not a valid event number to read from.

Instead, if there is NoStream, issue the read from 0