Authentication handles NotHandled message rather than waiting for timeout.
Now the authentication handles the NotHandled message which is generated when the node is not in a state to be able to repond to reads (such as while becoming leader after an election).

Prior to this change the authentication would not fail until it timed out 10s later, causing the client to be waiting an unnecessarily long time.

Also
- Handle timeouts via RequestResponseDispatcher so that it can tidy up the entry in its own map, avoiding leak.
- Pass the whole RequestResponseDispatcher into the schedule messages rather than allocating lots of closures that will end up in gen2