Add busy/idle metric for queues
- we time, using a stopwatch, how long each queue has been busy for
- when observed, we take the average busy time per queue group (whether groups are defined by the configuration)
- promql can then see how much the average time has changed since the begining of the period and calculate the percentage

note that we couldn't take the max instead of the average:
say we were measuring two queues in the group
and on scape 1 queue1 had been active for 10s, and queue2 had been active for 0s. The scrape would return 10s.
then on scrape 2, 5 seconds later, queue1 had been active for 10s still, but queue2 had been active for 5s. Well the max busyness for the group ought to be 100% but the scrape will return 10s which is the same as before making it seem like no work was done.

Taking the average avoids this problem and is equally (or more) valid to meaure because 100% means that all the queues are fully busy.