Make checkpoints always cached
Previously, checkpoints did happen to always be cached in the production code, but they had the capability to not be.

When cached, Reads and Writes are always through an Interlocked call which provides a full memory barrier, but when not cached the Read calls read through to the underlying storage (FileStream or MemoryMappedFile respectively), without a memory barrier.

The memory barrier can be important. For example when the chunk WriterWorkItem writes to the active chunk and it is cached, the writes are to some unmanaged memory, and then it moves the writer checkpoint. A thread reading the writer checkpoint (e.g. chaser, replication, etc) can then be sure that its reads from the unmanaged memory (via a chunk ReaderWorkItem) will be correct up to the checkpoint value that was read. This is because the barrier on write ensures that the data is written before the checkpoint is updated and the barrier on read ensures that the checkpoint is read before the data is read.

This does mean that these checkpoints cannot be used to communicate between processes and the tests have been updated to reflect that