Make an explicit lock for chunk Caching/Uncaching/TryDestructMemStreams
CacheInMemory and UnCacheFromMemory were correctly synchronized, but UnCacheFromMemory doesn't do the whole uncaching.
The uncaching continues in TryDestructMemStreams each time a mem reader is returned, and eventually the _cachedData is freed.
These continuations of uncaching can be concurrent with a subsequent call to CacheInMemory resulting in a variety of potential problems (elaborated in the PR)

This commit makes an explicit lock around all three parts of the caching process: CacheInMemory, UnCacheFromMemory and TryDestructMemStreams
so that the variables _cacheStatus, _cachedData, and _cachedLength are synchronized by the lock, and so is the creation and removal of the mem readers.

CacheInMemory and UnCacheFromMemory are called relatively infrequently, and won't contend with each other. We aren't worried about performance implications of the lock for those.
However, ReturnReaderWorkItem is called frequently on the read path. To help make this call quick we don't acquire the lock every time we return a memory work item,
instead we use a lock-free check to determine if we might need to acquire the lock first.
Then, if we do need to acquire the lock, it would be a shame if we got blocked waiting for a slow CacheInMemory call, but to minimise the risk of this, we make _isCached into _cacheStatus and give it three states. The chunk will only be recached after it has finished uncaching to returning mem readers should rarely, if ever, occur while caching.

I made new methods for the locks to make the diff small, I'll inline the locks in the next commit.

ObjectPool suffered from a similar problem, which we fix here also.