Remove reliance on UseNet5CompatFileStream
Co-authored-by: Timothy Coleman <timothy.coleman@gmail.com>

We used Net5CompatFileStream until now because in net6+ if one filestream writes to a file, another filestream that is reading from the same file cannot necessarily tell that anything has changed.

This change makes it so that FileStreams are not instantiated for reading until the chunk is readonly. Instead, the active chunk is always cached, and reads are served from the memstream.