Prevent replication checkpoint from moving incorrectly on follower
Before this change it was possible for a follower that needs truncating to expose the unreplicated data before truncating it:

0. follower has extra data from last time it was leader that will need to be truncated.
1. follower sends a request to subscribe to leader, but doesn't hear back quickly.
2. follower sends another request to subscribe to leader.
3. leader responds to both subscribe requests by sending the ReplicaSubscribed and LeaderReplicatedTo.
4. follower receives the first ReplicaSubscribed BUT discards it because it does not have the current subscription id.
   (if it had the current id we would have detected that truncation is required and stopped the replication tracking service)
5. follower receives the first LeaderReplicatedTo message and incorrectly updates its replication checkpoint to match.
    - this can result in unreplicated data being read or sent in subscriptions.
6. follower receives the second ReplicaSubscribed message (with the current subscription id) and stops the replication tracking service but it is too late.
7. follower receives the second LeaderReplicatedTo message.

The fix causes the leader to recognise that the follower will go offline for truncation and not sent it any more messages
(i.e. not the initial LeaderReplicatedTo message and no subsequent ones either).

Even if we implement online truncation later, the node can still resubscribe after it has done the truncation.

This involves no changes to the replication protocol.