Allow use of an Assembly that is already loaded into another context
Also sorted out context unloading in the tests on windows.
Long term I think a better approach might be to construct the resolver with the path to the plugin dll
so that it can resolve its dependencies, but this would require identifying which dll is the plugin dll
(perhaps by requiring the directory to have the same name) but also would require changing the way
our plugins are built.