Add support for multiple media types specified in one Accept header
Previously if you issued a GET to any of the non-grpc http endpoints with Accept = "text/plain;version=0.0.4;q=1,*/*;q=0.1" or more simply "text/plain,*/*"
then you would get a 406 Not Acceptable because the media types from the single header were not being parsed into multiple comma separated types.

We have also changed the metrics controller to support text/plain since that is the content type of the response it generates.

Unit test limitation: Unfortunately we can't figure out how to get the HTTP client to send a request that causes the problem, so the included test passes even before the fix.
Everything we tried seemed to result in a request that the server handled fine, perhaps it is automatically splitting the string into multiple headers when we set it.