Remove ConfigureAwait(false) calls and analyzer
I think these were originally added because there were some places in the code (mostly in the client and embedded client) where they there necessary.
These days the client lives in another repository and there is no embedded client. There aren't any synchronization contexts to forget in the server code and
none of the async paths are hot enough that we need the cycles saved by not checking for a sync context, so we can remove the ugly calls.