Tidy/standardize http pipeline
UseEndpoints registers endpoints. UseRouting selects an endpoint from those registered to that builder (app) instance.
The _first_ call to UseEndpoints is when the selected endpoint is executed.

For us this meant that the selected endpoint was executed on the UseEndPoints call on old line 142 (and not later as it might appear at first glance)

This is what necessitated the OPTIONS check on old line 137 _above_ the first UseEndPoints call (otherwise the gRPC code handles the OPTIONS request in UseEndPoints, and returns 404 if it is for a non gRPC endpoint)

Note that the non gRPC endpoints are registered in a _different_ builder because UseWhen (e.g. old line 143) gives you a different builder which needs its own calls to UseRounting and UseEndpoints. Therefore the original call to UseRouting on old line 136 does not know about the non grpc endpoints.

The idea in this PR is to make the main pipeline easier to understand by having only one call to UseEndpoints. It is then clear when the endpoint is executed and the gRPC code no longer intercepts OPTIONS requests that were intended for different endpoints.