Make metrics series work with OTLP exporter as well as Prometheus
This addresses two problems

1. newer versions of OpenTelemetry.Exporter.Prometheus.AspNetCore have
changed the names of the exported series (e.g. adding _totals to the end
of counters), breaking our dashboard.

for now we are reverting to an older version of the library.

2. OpenTelemetry.Exporter.Prometheus.AspNetCore appends the units to the
metrics names but OpenTelemetry.Exporter.OpenTelemetryProtocol does not,
meaning the latter will not work with our dashboard.

it's possible to apply transformations in the OTEL collector but we
avoid this so that users don't have to update their collector
configuration when we add more metrics.

instead for now we are putting the units directly in the metric names.
this causes both exporters to export the same series.

this feels a bit strange and i may well be misunderstanding something
here. we will revisit when we (re)upgrade the otel library versions.