Fixed: Redacted events break replication on follower nodes
During replication, a follower node deserializes events into 'PrepareLogRecord's from the TCP stream (in memory) then writes the 'PrepareLogRecord's to disk.
Since the `Data` property of a redacted event is set to an empty byte array, follower nodes would write empty data to disk instead of the redacted data.
This would cause the follower node to crash as the writer stream would not be at the expected position.