Bump up the chunk format version to 4 (`Transformed`)
v4 basically represents the addition of the `TransformType` field.
We want to use v4 for all new chunks, whether a transformation is applied or not (i.e if we're using the `Identity` transform).

However, the existing chunk version verification code throws an exception on startup whenever a chunk with an unknown version is opened.
This means that old versions of the database will not be able to read/write to v4 chunks even if no transformation is applied.

To work around this, we implement forward compatibility in the following way:
- the current `Version` field is renamed to `MinCompatibleVersion`
- the next free slot of the chunk header is used to store the actual version (`Version`)

`MinCompatibleVersion` represents the minimum chunk version that the db needs to understand to be able to read/write to this chunk.
If a new chunk format is forward compatible, then `MinCompatibleVersion` can still point to an older version.
However, if a change is not forward compatible (e.g when a transform is applied), then `MinCompatibleVersion` must also be bumped up to a higher value.