Update streams.md
Removed the misleading statement, 

Regarding deleted streams retaining one event that is readable, 

As per slack chat with Alexey that seems to not be an accurate representation. 

It seems there is enough additional information regarding the behavior of soft delete where appended events to a soft deleted streams start at previous offset plus one, that the best option is to simply delete this paragraph. 

At which version this paragraph was true? I don't have enough information to make that call, but certainly it is NOT true for current version.