Log SLOW QUEUE messages by default again
When EventStore:Metrics:Queues:Processing is false (which is it by default), the tracker which is used was not returning the current time when recording.
This caused the queues to think that no time had passed while processing the message and prevented slow queue messages from being logged.

Slow BUS messages were still logged