Skip verifying chunk hashes when remote
Because it is expensive, and remote storage such as S3 is rather robust.
In the future we could make this configurable, perhaps dependent on the storage implementation

We also log warnings when reading the header/footer of a chunk outside of TFChunk itself.
These methods are called on lots of chunks at once (e.g. on startup) but ought not to be called
for many remote chunks. We want to be warned about it because it will be slow.