Validate against attempts to set metadata for the "" stream (#4799)
"" has never been a valid stream name. Attempting to set the metadata for it results in an attempt to write to the stream "$$" which, until now, has been a valid stream name.

However, writing to $$ involves checking on the "" stream, to see if it is soft deleted. This results in the storage writer exiting which shuts down the server to avoid a 'sick but not dead' scenario

This PR makes "$$" an invalid stream name, and so the attempt to write to it is rejected at an early stage.