Track initialized services so we know what to wait for on shutdown
Before this commit, on shutdown we waited for a hard coded number of services to terminate but
1. The count of services was wrong (didn't include ReplicationTrackingService)
2. Not all unit tests that start the server actually involve starting all the services (e.g. starting a single node of a three node cluster)
   in which case shutdown will timeout waiting for services to terminate that never started.

Now we ensure that any service that sends ServiceShutdown also sends ServiceInitialized, and we track which services were initialized to know
which services we need to wait for on shutdown.

The overall effect here is that the server should shutdown smoothly in more scenarios.