Refactor the chunk switching so that it not done via the ChunkWriterForExecutor
Instead the ChunkWriterForExecutor passes the completed chunk back to the main ChunkExecutor and it does the swap. The swap is still done via the ChunkManager.

This is preparation for scavenging the archive, where we will want to upload the newly scavenged chunk instead of swapping it in.

Also added a trace when aborting chunk writing