Prevent swallowing of gRPC exceptions that lead to hanging writes on client (#4857)
Reverting f9791cf - Reduce gRPC log level on the server

If we swallow the exception the call terminates successfully, which the client isn't expecting in case of failure. e.g. in the dotnet client the user's Append call will not be terminated.

There is room to discuss later if we should be converting some of these to RpcExceptions, to avoid logging Errors for them.