Always cache midpoints for scavenged chunks (#4897)
Previously we cached midpoints exactly for chunks that are not in the chunk cache.
This change caches midpoints even for chunks that are in the chunk cache.

The size of the midpoints array is only about 12k bytes per chunk, so adds little overhead on top of the chunk itself.

The advantage is that it saves us reading the whole posmap table into a buffer (which is what TranslateExactPosition/TranslateClosestForwardPosition currently does without midpoints)
or repeatedly repositioning the readerworkitem (which is what the same methods used to do in v24.10). We prefer not to revert to the old v24.10 behaviour because
using the midpoints is still good for performance even when cached, and especially so when encryption-at-rest is in use.

When populating the midpoints we also now return [] rather than null when the chunk is empty. This avoids reacquiring the lock next time.